Add Image class and current reconstruction infrastructure
The Image class takes a 2D matrix of data, keeps track of the image
size, pixel size, and number of pixels, and offers methods to resize,
resample, deconvolve a point spread function, etc.

Right now it also has the capability do current inversion (assuming
it’s a flux image), but it would make sense to make a FluxImage class
that does things particular to flux images. Just to be general.