Integer dictionary keys saved and loaded properly
jsonpickle and/or HDF5 did not like dictionary keys that are ints. The
integer keys are converted to strings upon saving and converted back to
ints upon loading. Why not just use an array? Dictionaries add a certain
extra level of convenience, and make it clear that the data sets are
separate.