[dhl88] small changes with gmf57 to make scanning in DR work, reorganized touchdown
Small edits to attocube, piezo, squidarray to correct errors
when initially scanning in DR.

Did not understand touchdown at all, so I create helper functions
and moved the code around to make it easier to read.

Added ability to change the wait time when waiting for the
capacitance to settle.  Ended up not mattering.  Changing the
filter settings is all that mattered.