Add function to calculate proper extents for imshow
Originally, was using just the minimum and maximum of x and y arrays,
but this does not correctly position the extrema at the center of the
boundary pixels. We have to add and subtract offsets to center the
boundary pixels at the extrema of x and y arrays.

Implemented in planefit, scanplane, and general plot2D method in
plot_mpl.